refactor: LottoSystem클래스에 컨트롤러랑 비즈니스 로직 분리하여 컨트롤러 클래스 생성
